Refactor paymaster client to the paymaster client utility file
