+++ template/web/app/paymaster-bundler/api/PaymasterProxy.ts
@@ -3,18 +3,7 @@ import { paymasterActionsEip7677 } from 'permissionless/experimental';
 import { createClient, createPublicClient, http } from 'viem';
 import { baseSepolia } from 'viem/chains';
 import { willSponsor } from '../utils';
-
-const paymasterService = process.env.NEXT_PUBLIC_PAYMASTER_URL ?? '';
-
-export const client = createPublicClient({
-  chain: baseSepolia,
-  transport: http(),
-});
-
-export const paymasterClient = createClient({
-  chain: baseSepolia,
-  transport: http(paymasterService),
-}).extend(paymasterActionsEip7677({ entryPoint: ENTRYPOINT_ADDRESS_V06 }));
+import { paymasterClient } from '../../../../web/src/components/SmartWallets/PaymasterClient';
 
 type PaymasterRequest = {
   method: string;
