+++ template/web/src/components/SmartWallets/PaymasterClient.ts
@@ -1,0 +1,16 @@
+import { createClient, createPublicClient, http } from 'viem';
+import { baseSepolia } from 'viem/chains';
+import { ENTRYPOINT_ADDRESS_V06 } from 'permissionless';
+import { paymasterActionsEip7677 } from 'permissionless/experimental';
+
+const paymasterService = process.env.NEXT_PUBLIC_PAYMASTER_URL ?? '';
+
+export const client = createPublicClient({
+  chain: baseSepolia,
+  transport: http(),
+});
+
+export const paymasterClient = createClient({
+  chain: baseSepolia,
+  transport: http(paymasterService),
+}).extend(paymasterActionsEip7677({ entryPoint: ENTRYPOINT_ADDRESS_V06 }));
